[BPK-4276] Restructure Android pages
