+++ docs/src/pages/OverlayPage/OverlayPage.js
@@ -22,9 +22,9 @@ import React from 'react';
 
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
-import Android from '../AndroidOverlayPage';
 import Web from '../WebOverlayPage';
 
+import Android from './AndroidOverlayPage';
 import IOS from './IOSOverlayPage';
 
 const Page = () => (
