+++ docs/src/pages/PanelPage/AndroidPanelPage.js
