+++ docs/src/pages/PanelPage/PanelPage.js
@@ -22,9 +22,9 @@ import React from 'react';
 
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
-import Android from '../AndroidPanelPage';
 import Web from '../WebPanelsPage';
 
+import Android from './AndroidPanelPage';
 import IOS from './IOSPanelPage';
 import Native from './NativePanelsPage';
