+++ docs/src/pages/RatingPage/AndroidRatingPage.js
