+++ docs/src/pages/SnackbarPage/SnackbarPage.js
@@ -22,8 +22,8 @@ import React from 'react';
 
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
-import Android from '../AndroidSnackbarPage';
 
+import Android from './AndroidSnackbarPage';
 import IOS from './IOSSnackbarPage';
 import Native from './NativeSnackbarPage';
