+++ docs/src/pages/SpinnerPage/AndroidSpinnerPage.js
