+++ docs/src/pages/SpinnerPage/SpinnerPage.js
@@ -22,9 +22,9 @@ import React from 'react';
 
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
-import Android from '../AndroidSpinnerPage';
 import Web from '../WebSpinnersPage';
 
+import Android from './AndroidSpinnerPage';
 import IOS from './IOSSpinnerPage';
 import Native from './NativeSpinnerPage';
