+++ docs/src/pages/StarRatingPage/AndroidStarRatingPage.js
