+++ docs/src/pages/SwitchPage/AndroidSwitchPage.js
