+++ docs/src/pages/TextInputPage/TextInputPage.js
@@ -21,8 +21,8 @@ import React from 'react';
 import IntroBlurb from '../../components/IntroBlurb';
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import Web from '../WebTextInputPage';
-import Android from '../AndroidTextFieldPage';
 
+import Android from './AndroidTextFieldPage';
 import IOS from './IOSTextInputPage';
 import Native from './NativeTextInputPage';
