+++ docs/src/pages/TextPage/AndroidTextPage.js
