+++ docs/src/pages/TextPage/TextPage.js
@@ -22,9 +22,9 @@ import React from 'react';
 
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
-import Android from '../AndroidTextPage';
 import Web from '../WebTextPage';
 
+import Android from './AndroidTextPage';
 import IOS from './IOSTextPage';
 import Native from './NativeTextPage';
